Adding areas to claim planets.
The new idea is to have areas for different teams, and each team has to grapple the planets into an area, and keep them in there for the rest of the game. Each team will be able to nick each others planets, and once the game time is up the total mass in each area is counted up, and the team with the most mass wins.